Move logger to logger package
